test suite fixes for some locales; fix to make bind -x bindings work with negative arguments; several portability fixes from pkgsrc; fixes for funsubs with return builtin state and local OPTIND

+++ lib/malloc/malloc.c
@@ -299,7 +299,7 @@ static const unsigned long binsizes[NBUCKETS] = {
 };
 
 /* binsizes[x] == (1 << ((x) + 5)) */
-#define binsize(x)	binsizes[(x)]
+#define binsize(x)	binsizes[(size_t) (x)]
 
 #define MAXALLOC_SIZE	binsizes[NBUCKETS-1]
